+++ example.ts
@@ -3,15 +3,23 @@ import { Changelog, Release } from "./mod.ts";
 
 const changelog = new Changelog("Changelog")
   .addRelease(
-    new Release("2.6.0", "2024-07-21")
+    new Release("2.6.0", "2025-02-22")
       .added(
         "New option `--no-v-prefix` to generate the tag names without prepending `v` #43",
       )
       .added(
         "New option `--no-sort-releases` to disable sorting of releases #51",
       )
-      .added("`--help, -h` flag [#45].")
-      .fixed("Updated dependencies"),
+      .added(
+        "New option `--bullet-style` for customizing the bullet point character used in unordered lists #48",
+      )
+      .added("Add Azure DevOps URL support #49")
+      .added(
+        "New option `--combine` to combine changes with the same version #50",
+      )
+      .added("`--help, -h` flag #45.")
+      .fixed("Updated dependencies")
+      .fixed("Skip release creation if it already exists #47"),
   )
   .addRelease(
     new Release("2.5.3", "2023-11-19")